pyproject.toml
[project]
name = "animal-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "denisebozza",email = "denisebozza01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pylint (>=3.3.7,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "kagglehub (>=0.3.12,<0.4.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "torch (>=2.7.1,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"